Add files via upload
This file solved the cache problem related to the binder. The problem was causing outdated results on the binder page when the user went from the binder page to another page and then came back. On the other hand, when the binder products were removed/added new ones.

The solution was to add this script that basically erases the cache just for the Binder page.